feat: Add search functionality and update button style
This commit modifies the HTML structure to add a search button and functionality to the search bar. It also updates the style of the search button to use a material symbol outlined icon. The changes in the main.js file include fixing a typo in the id of the favorites button.

Resolves: #123